+++ changelog.md
@@ -8,6 +8,7 @@ Codion Change Log
 - is.codion.plugin.swing.mcp module added.
 - SwingMcpServer, unnecessary title bar click removed, refactoring.
 - SwingMcpServer.takeApplicationScreenshot() reimplemented, bypassing the robot, window no longer has to be in front.
+- SwingMcpServer, window listing tool added and window screenshot functionality improved, refactoring.
 ### is.codion.swing.common.ui
 - FilterTable.Builder.sortingEnabled() renamed sortable().